FLEXLEXER_PATH never worked -- fixed.
This fix concerns old mac os systems and new freebsd in first place.
The bug appeared in case when flex (with FlexLexer.h) is installed in
system but is too old. In that case we could give a path to new flex
but compiler never finds correcponding FlexLexer.h and uses old one =>
nothing compiled in specific versions of flex.

The fix may be used somewhere else, if we have similar scenarios for
libraries search.